+++ src/apis/problem.ts
@@ -1,4 +1,4 @@
-import { axiosInstance } from './axiosInstance'
+import { axiosInstance } from './axios-instance'
 import {
     GetProblemDetailProps,
     GetProblemListProps,
