+++ src/constants/difficulty-data.ts
