+++ src/constants/sector-data.ts
